+++ test/unit/client.test.ts
@@ -319,7 +319,7 @@ describe('Client', () => {
   describe('fetchUsersStation()', () => {
     let station: IStation;
     let user: IPersonDetails;
-    let mockStationResponse: Promise<{ body: IStation }>;
+    let mockStationResponse: Promise<{ data: IStation }>;
 
     beforeEach(() => {
       const stationId = 'the-bat-phone';
@@ -346,8 +346,8 @@ describe('Client', () => {
       };
       jest.spyOn(utils, 'requestApiWithRetry').mockImplementation(() => {
         return {
-          promise: mockStationResponse || Promise.resolve({ body: station })
-        } as RetryPromise<{ body: IStation }>;
+          promise: mockStationResponse || Promise.resolve({ data: station })
+        } as RetryPromise<{ data: IStation }>;
       });
       sdk = constructSdk();
     });
@@ -835,7 +835,7 @@ describe('Client', () => {
       const session2 = new MockSession();
 
       sessionManagerMock.getAllJingleSessions.mockReturnValue([session1, session2] as any);
-      sessionManagerMock.endSession.mockResolvedValue(null);
+      sessionManagerMock.forceTerminateSession.mockResolvedValue(null);
       mediaMock.destroy.mockReturnValue();
 
       jest.spyOn(sdk, 'removeAllListeners');
@@ -849,8 +849,8 @@ describe('Client', () => {
           { sessionId: session2.id, conversationId: session2.conversationId },
         ]
       });
-      expect(sdk.sessionManager.endSession).toHaveBeenCalledWith(session1);
-      expect(sdk.sessionManager.endSession).toHaveBeenCalledWith(session2);
+      expect(sdk.sessionManager.forceTerminateSession).toHaveBeenCalledWith(session1.id);
+      expect(sdk.sessionManager.forceTerminateSession).toHaveBeenCalledWith(session2.id);
       expect(sdk.removeAllListeners).toHaveBeenCalled();
       expect(sdk.media.destroy).toHaveBeenCalled();
       expect(sdk.disconnect).toHaveBeenCalled();